pyproject.toml
[project]
name = "crossd_metrics"
version = "0.1.0"
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "gql>=3.4.1",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.2",
    "rich>=13.4.2",
    "PyGithub>=1.59.0",
    "requests-toolbelt>=1.0.0",
    "html5lib>=1.1",
    "GitPython>=3.1.44",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
